Check before doing things

+++ chrome/pack.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-rm -r de-t-co/assets
+if [ -d "de-t-co/assets" ]; then rm -r de-t-co/assets; fi
 
 mkdir -p de-t-co/assets
 
@@ -8,6 +8,6 @@ convert de-t-co.svg -flatten -resize 128x128 de-t-co/assets/icon128.png
 convert de-t-co.svg -flatten -resize 48x48 de-t-co/assets/icon48.png
 convert de-t-co.svg -flatten -resize 16x16 de-t-co/assets/icon16.png
 
-rm de-t-co.zip
+if [ -e "de-t-co.zip" ]; then rm de-t-co.zip; fi
 cd de-t-co
 zip ../de-t-co.zip src/* assets/* manifest.json
